Set prayer times correctly and also set method for prayer times and make it in Prefrences

+++ app/src/main/java/com/MohamedTaha/Imagine/New/rest/APIServices.java
@@ -11,13 +11,11 @@ public interface APIServices {
     @GET("calendar")
     Call<Azan> getPrayerTimes(@Query("latitude") double latitude,
                               @Query("longitude") double longitude,
-                              @Query("annual") boolean annual, @Query("method") double  method);
-
+                              @Query("annual") boolean annual, @Query("method") int  method);
     @GET("calendarByCity")
     Call<Azan> getPrayerTimesByCity(@Query("city") String city,
-                                    @Query("country") String country,
-                                    @Query("method") double  method);
-
+                                     @Query("country") String country,
+                                     @Query("method") int  method);
     @GET("json")
     Call<GetCity> getCity();
 }
